requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
distro==1.9.0
emoji==2.10.1
environs==9.5.0
exceptiongroup==1.2.0
filelock==3.13.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.2.0
fuzzywuzzy==0.18.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
InstructorEmbedding==1.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.1.2
langchain-community==0.0.17
langchain-core==0.1.18
langdetect==1.0.9
langsmith==0.0.83
Levenshtein==0.24.0
libmagic==1.0
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.20.2
minio==7.2.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.11.1
packaging==23.2
pandas==2.2.0
pg8000==1.30.4
pillow==10.2.0
protobuf==4.25.2
pyarrow==15.0.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
pymilvus==2.3.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.1.2
python-Levenshtein==0.24.0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
scramp==1.4.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.37.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
unstructured==0.12.3
unstructured-client==0.17.0
urllib3==2.2.0
wrapt==1.16.0
yarl==1.9.4
